config/requirements.txt
Flask==3.1.1
simple-salesforce==1.12.6
python-dotenv==1.0.1
openai>=1.90.0
openpyxl==3.1.5
pandas==2.2.3 